+++ CHANGELOG.md
@@ -18,6 +18,7 @@ SHA: ``8201917e3f9b47b8edd8039ea3278ef8631b0f2a``
   * Added support for the L-BFGS-B optimizer. (#296)
   * Added GPU implementation for q,p-EI and its gradient computation. (#219)
     * Speed up GPU functions by redesign of memory allocation. (#297)
+    * Connected GPU functions to multistart gradient descent optimizer. (#270)
 
 * Changes